requirements.txt
mpi4py
pyyaml
numpy
matplotlib
torch
torchvision
